Refactor CredentialsSection
